Add tests (#6)
Added a simple test suite supported by `testcode`, following the example of Quantum ESPRESSO.

The test suite involves a PBE calculation, followed by a PZ innerloop calculation, followed by a KI calculation, all on ozone. The test checks the total energy, the ODD energy, the HOMO, and the LUMO.

The tests are run via `make check`, and are run during the checks for this repo.