Storing ODD potentials in reciprocal space (#13)
This PR changes how we store the potential `vsic` (and related quantities) from on the real-space grid (dimension `nnrx`) to the G-vector grid (dimension `ngm`). This should reduce the peak memory consumption (because `vsic` is smaller) but increase slightly the computational cost (because we must perform additional FFTs to swap back and forth from real to reciprocal space where required).

Additionally, in `nksic_rot_emin_cg_general` there were three copies of the potential (`vsic`, `vsic1`, and `vsic2`). This was the most memory-intensive part of the calculation. I was able to remove one of these copies of the potential entirely.

For my simple test systems, these changes collectively reduced the peak memory demand by ~ 30% and increased the compute time by ~ 10%.

The `nk` and `nkpz` corrections have been removed (because I did not bother implementing the G-vector change for them). The other corrections have been split off into a new module, `nksic_corrections`, to avoid the `interface` at the start of `nksiclib`.