Tweak phasefactor implementation so that it does not trigger an ifx bug (#24)
The new intel compiler `ifx` was producing `NaN`s for the simplest of calculations. This was traced back to the function `phfacs_x` in `phasefactor.f90`.

Any of the following changes fix the bug
1.  compiling with `O1` or lower
2.  compiling with `-check bounds`
3. adding `write` statements to the code
4. refactoring the calculation of `ei1`-`ei3`

We do not fully understand the precise nature of the bug. Points 2 and 3 indicate that vectorisation is the issue, since the main difference between `O1` and `O2` is autovectorisation, and `write` statements disable vectorisation, but using `-qopt-report` suggested that in the old implementation the loops weren't being vectorised...

We will potentially submit a bug report to intel, but for the moment we will adopt the refactor as an easy way of avoiding the issue.